intToLong

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[]{1,10,0};
+        long[] sales = new long[]{1,10,0};
         SalesManager manager = new SalesManager(sales);
         System.out.println(manager.max());
     }
